--- app/src/main/res/drawable/selector_bottom_my_icon.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:state_selected="false"
-        android:drawable="@drawable/my" />
-    <item
-        android:state_selected="true"
-        android:drawable="@drawable/my_act" />
-</selector>
