--- app/src/main/res/drawable/selector_bottom_nav_folder_icon.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:state_selected="false"
-        android:drawable="@drawable/folder" />
-    <item
-        android:state_selected="true"
-        android:drawable="@drawable/folder_act" />
-</selector>
