--- app/src/main/res/drawable/selector_bottom_nav_noti_icon.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:state_selected="false"
-        android:drawable="@drawable/noti" />
-    <item
-        android:state_selected="true"
-        android:drawable="@drawable/noti_act" />
-</selector>
